refactor: remove the old loro-js folder
