+++ crates/loro-wasm/tsconfig.json
@@ -27,7 +27,7 @@
         /* Modules */
         "module": "ESNext" /* Specify what module code is generated. */,
         // "rootDir": "./",                                  /* Specify the root folder within your source files. */
-        // "moduleResolution": "node10",                     /* Specify how TypeScript looks up a file from a given module specifier. */
+        "moduleResolution": "node", /* Specify how TypeScript looks up a file from a given module specifier. */
         // "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
         // "paths": {},                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
         // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
@@ -53,7 +53,7 @@
         // "sourceMap": true,                                /* Create source map files for emitted JavaScript files. */
         "inlineSourceMap": true, /* Include sourcemap files inside the emitted JavaScript. */
         // "outFile": "./",                                  /* Specify a file that bundles all outputs into one JavaScript file. If 'declaration' is true, also designates a file that bundles all .d.ts output. */
-        "outDir": "./dist" /* Specify an output folder for all emitted files. */,
+        // "outDir": "./dist" /* Specify an output folder for all emitted files. */,
         // "removeComments": true,                           /* Disable emitting comments. */
         "noEmit": false /* Disable emitting files from a compilation. */,
         // "importHelpers": true,                            /* Allow importing helper functions from tslib once per project, instead of including them per-file. */
@@ -107,10 +107,10 @@
         }
     },
     "exclude": [
-        "node_modules",
         "dist",
         "deno",
         "scripts",
-        "deno_tests"
+        "deno_tests",
+        "node_modules"
     ]
 }
