+++ crates/loro-wasm/vite.config.ts
@@ -6,7 +6,8 @@ export default defineConfig({
   test: {
     exclude: [
       ...configDefaults.exclude,
-      "deno/*"
+      "deno/*",
+      "deno_tests/*"
     ]
   },
 });
