--- loro-js-tests/.npmignore
@@ -1,10 +1,0 @@
-tests/
-.vscode/
-deno/
-src/
-node_modules/
-pnpm-lock.yaml
-rollup.config.mjs
-tsconfig.json
-vite.config.ts
-CHANGELOG.md
