--- loro-js-tests/.vscode/settings.json
@@ -1,8 +1,0 @@
-{
-  "deno.enable": false,
-  "editor.defaultFormatter": "vscode.typescript-language-features",
-  "editor.formatOnSave": true,
-  "[typescript]": {
-    "editor.defaultFormatter": "vscode.typescript-language-features"
-  }
-}
