--- loro-js-tests/deno/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-  "deno.enable": true
-}
