--- loro-js-tests/deno/deno.json
@@ -1,5 +1,0 @@
-{
-  "name": "@loro/loro",
-  "version": "0.1.0",
-  "exports": "./mod.ts"
-}
