--- loro-js-tests/deno/mod.ts
@@ -1,1 +1,0 @@
-export * from "npm:loro-crdt@0.14.2";
