--- loro-js-tests/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "loro-crdt-test",
-  "version": "1.0.7",
-  "description": "Loro CRDTs is a high-performance CRDT framework that makes your app state synchronized, collaborative and maintainable effortlessly.",
-  "keywords": [
-    "crdt",
-    "CRDTs",
-    "realtime",
-    "collaboration",
-    "sync",
-    "p2p"
-  ],
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/loro-dev/loro.git"
-  },
-  "main": "dist/loro.js",
-  "module": "dist/loro.mjs",
-  "typings": "dist/loro.d.ts",
-  "scripts": {
-    "build": "rollup -c",
-    "watch": "rollup -c -w",
-    "test": "node --expose-gc ./node_modules/vitest/vitest.mjs run && npx tsc --noEmit",
-    "prepublish": "pnpm run build"
-  },
-  "author": "Loro",
-  "homepage": "https://loro.dev",
-  "license": "MIT",
-  "dependencies": {
-    "loro-wasm": "workspace:*"
-  },
-  "devDependencies": {
-    "@rollup/plugin-node-resolve": "^15.0.1",
-    "@typescript-eslint/parser": "^6.2.0",
-    "@vitest/ui": "^1.0.4",
-    "esbuild": "^0.18.20",
-    "eslint": "^8.46.0",
-    "loro-crdt-old": "npm:loro-crdt@=0.16.0",
-    "loro-crdt-alpha-4": "npm:loro-crdt@=1.0.0-alpha.4",
-    "prettier": "^3.0.0",
-    "rollup": "^3.20.1",
-    "rollup-plugin-dts": "^5.3.0",
-    "rollup-plugin-esbuild": "^5.0.0",
-    "typescript": "^5.0.2",
-    "vite": "^4.2.1",
-    "vite-plugin-wasm": "^3.2.2",
-    "vitest": "^1.4.0"
-  }
-}
