+++ package.json
@@ -9,9 +9,9 @@
     "build": "cargo build",
     "test": "cargo nextest run --features=test_utils,jsonpath --no-fail-fast && cargo test --doc",
     "test-all": "pnpm test && pnpm test-wasm",
-    "test-wasm": "cd crates/loro-wasm && deno task build-dev && cd ../../loro-js && pnpm i && pnpm run test",
+    "test-wasm": "cd crates/loro-wasm && pnpm i && pnpm build-dev",
     "coverage": "mkdir -p coverage && cargo llvm-cov nextest --features test_utils,jsonpath --lcov > coverage/lcov-nextest.info && cargo llvm-cov report",
-    "release-wasm": "cd crates/loro-wasm && deno task build-release && cd ../../loro-js && pnpm i && pnpm build && pnpm run test",
+    "release-wasm": "cd crates/loro-wasm && pnpm i && pnpm build-release",
     "check": "cargo clippy --all-features -- -Dwarnings",
     "run-fuzz-corpus": "cd crates/fuzz && cargo +nightly fuzz run all -- -max_total_time=1",
     "fix": "cargo clippy --fix --features=test_utils",
